wedding-face-album-app/backend/requirements.txt
fastapi
uvicorn
python-multipart
face_recognition
numpy
Pillow